requirements.txt
graphql-core-next
uvicorn
starlette
